Add a licenseRule.json file for checking of SPDX license identifiers
The license test in qtqa reads this file in order to check the validity
of the license used.
This file reproduce the QUIP-18 [1] rules, with some exceptions.

Each entry in the file corresponds to a set of licensing rules.

A set of licensing rules can depend on the file ending, registered
in "file_pattern_ending". The last entry in the file has no
"file_pattern_ending", it sets the rules for the files whose
licensing does not depend on their ending.

The license to be used depends on the location of the file within
the Qt module repository. Let's call this "<true_location>". The
"<true_location>" can also correspond to a file name, offering
flexibility for exceptions to the rule. The "<true_location>" are
registered in "location".

For each "<true_location>" there is a "file type" entry and a "spdx"
entry. The "spdx" entry gives the rule: the expected license tag(s) in
SPDX format for the file ending (if applicable) and "<true_location>".
The "file type" informs on the QUIP-18 type the tested file corresponds
to. It is purely informational for the reader.

The set of rules are tested in order of appearance in the json file. For
this reason, a more constraining ending (like "special.txt") needs to
appear in a "file_pattern_ending" located before the
"file_pattern_ending" of a less constraining ending (like ".txt").
Also, a file ending cannot be present in two "file_pattern_ending".

"file_pattern_ending" and "spdx" should list strings.
"<true_location>" can be regular expressions.

During the test the deeper "<true_location>" are checked first. The
order is which they appear in the json file does not matter.

To test this file, run
QT_MODULE_TO_TEST=../qttranslations perl tests/prebuild/license/tst_licenses.pl

[1]: https://contribute.qt-project.org/quips/18

Task-number: QTBUG-121039
Change-Id: I4423ae5159c49f69a73e918e935c5dc4db88c248
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>