CMake: Avoid running xmllint on ts files on macOS hosts
Running xmllint on ts files to extract the source language is only
needed in the context of macOS bundle applications created by
qt_add_executable().

It is not necessary to extract the source languages for the Qt
translations, because we don't create macOS apps in the repo.

The extra process executions also slow down the configuration of the
repo, due to 280+ invocations of xmllint.

Set the QT_SKIP_STORE_LANGUAGES_FROM_TS_FILES variable to avoid the
needless processing.

Change-Id: I22696f0ca77f4b1f8dcf128cb2b5660b5a6f360b
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>