CMake: Let the 'ts-qtfoo-all' targets update all languages in one go
A ts-qtfoo-all target merely depended on the targets ts-qtfoo-de,
ts-qtfoo-fr, ts-qtfoo-fi, ... and so on, meaning lupdate would run once
for each language. That's very inefficient.

Now, ts-qtfoo-all contains the command to run lupdate on all .ts files
of qtfoo in one go. The ts-all target depends on all the ts-qtfoo-all
targets.

Fixes: QTBUG-130699
Change-Id: I3d95d8e9578eb547b9253f1b13ae5bd23394afbe
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>